Move sortNodesByExecutionOrder to app utils
