+++ packages/frontend/editor-ui/src/app/utils/workflowUtils.ts
@@ -1,5 +1,6 @@
 import type { IWorkflowDb, INodeUi } from '@/Interface';
 import type { ITag } from '@n8n/rest-api-client/api/tags';
+import type { IConnections } from 'n8n-workflow';
 
 /**
  * Converts workflow tags from ITag[] (API response format) to string[] (store format)
@@ -38,4 +39,66 @@ export function removeWorkflowExecutionData(
 	};
 
 	return sanitizedWorkflow;
+}
+
+interface ExecutionOrderItem {
+	node: { name: string; position: [number, number] };
+	isTrigger: boolean;
+}
+
+/**
+ * Orders nodes by execution order, grouped by trigger.
+ * Iterates triggers (sorted by X position), DFS-ing each trigger's subgraph
+ * to collect downstream nodes in execution order (depth-first, matching the
+ * backend v1 execution strategy). This lets users complete one full branch
+ * before moving to the next. Nodes reachable from multiple triggers appear
+ * only under the first trigger visited.
+ * Orphaned nodes (not reachable from any trigger) are dropped.
+ * When there are no triggers, returns an empty array.
+ */
+export function sortNodesByExecutionOrder<T extends ExecutionOrderItem>(
+	nodes: T[],
+	connectionsBySourceNode: IConnections,
+): T[] {
+	const triggers = nodes
+		.filter((item) => item.isTrigger)
+		.sort((a, b) => a.node.position[0] - b.node.position[0]);
+
+	if (triggers.length === 0) return [];
+
+	const itemsByName = new Map<string, T>();
+	for (const item of nodes) {
+		itemsByName.set(item.node.name, item);
+	}
+
+	const result: T[] = [];
+	const visited = new Set<string>();
+
+	for (const trigger of triggers) {
+		if (visited.has(trigger.node.name)) continue;
+		visited.add(trigger.node.name);
+		result.push(trigger);
+
+		// DFS through all workflow connections from this trigger
+		const dfs = (name: string) => {
+			const nodeConns = connectionsBySourceNode[name];
+			if (!nodeConns) return;
+			for (const type of Object.keys(nodeConns)) {
+				for (const outputs of nodeConns[type]) {
+					for (const conn of outputs ?? []) {
+						if (visited.has(conn.node)) continue;
+						visited.add(conn.node);
+						const item = itemsByName.get(conn.node);
+						if (item) {
+							result.push(item);
+						}
+						dfs(conn.node);
+					}
+				}
+			}
+		};
+		dfs(trigger.node.name);
+	}
+
+	return result;
 }
